third_party/blink/web_tests/external/wpt/css/selectors/focus-visible-018-expected.txt

This is a testharness.js-based test.
PASS ":focus-visible" should be a valid selector
FAIL Mouse focus does not show a focus ring by default assert_equals: outline-style for DIV#target should be none expected "none" but got "auto"
Harness: the test ran to completion.
